Delete push-notifications.txt
Switch to use uBO built-in EasyList – Annoyances ==> EasyList – Notifications 

By the way in the previous commit I blamed github for being unable to delete this file, but it has turned out it wasn't github's fault, it was EasyList – Annoyances ==> EasyList – Notifications causing it, which I enabled before deleting the old file, now once I've disabled the list, I can delete the file without problems, it's an irony: deleting the old filter list file to switch to uBO built-in list, just to realise that the built-in list prevents deletion of the old filter list file. Wow.